+++ src/commands/build.ts
@@ -9,7 +9,7 @@ import { TsrpcConfig } from "../models/TsrpcConfig";
 import { error } from "../models/util";
 import { genApiFilesByProto } from "./api";
 import { ensureSymlinks } from "./link";
-import { copyDirReadonly } from "./sync";
+import { copyDirReadonly, syncByConfig } from "./sync";
 
 export type CmdBuildOptions = {
     config?: TsrpcConfig
@@ -47,20 +47,7 @@ export async function cmdBuild(options: CmdBuildOptions) {
         // Auto Sync
         if (autoSync && options.config.sync) {
             const logger = options.config.verbose ? console : undefined;
-            // Copy
-            for (let confItem of options.config.sync) {
-                if (confItem.type === 'copy') {
-                    CliUtil.doing(`${i18n.copy} '${confItem.from}' -> '${confItem.to}'`);
-                    await copyDirReadonly(confItem.from, confItem.to, !!confItem.clean, logger);
-                    CliUtil.done(true);
-                }
-            }
-
-            // Symlink
-            await ensureSymlinks(options.config.sync.filter(v => v.type === 'symlink').map(v => ({
-                src: v.from,
-                dst: v.to
-            })), logger);
+            await syncByConfig(options.config.sync, logger);
         }
     }