+++ src/commands/sync.ts
@@ -23,22 +23,7 @@ export async function cmdSync(options: CmdSyncOptions) {
             return;
         }
 
-        const logger = options.config.verbose ? console : undefined;
-        // Copy
-        for (let item of options.config.sync) {
-            if (item.type === 'copy') {
-                CliUtil.doing(`${i18n.copy} '${item.from}' -> '${item.to}'`);
-                await copyDirReadonly(item.from, item.to, !!item.clean, logger);
-                CliUtil.done(true);
-            }
-        }
-
-        // Symlinks
-        await ensureSymlinks(options.config.sync.filter(v => v.type === 'symlink').map(v => ({
-            src: v.from,
-            dst: v.to
-        })), console);
-
+        await syncByConfig(options.config.sync, options.config.verbose ? console : undefined);
         console.log(chalk.green(i18n.allSyncedSucc))
     }
     else {
@@ -58,6 +43,27 @@ export async function cmdSync(options: CmdSyncOptions) {
         CliUtil.done(true);
         console.log(chalk.green(i18n.syncedSucc))
     }
+}
+
+export async function syncByConfig(syncConfig: NonNullable<TsrpcConfig['sync']>, logger: Logger | undefined) {
+    if (!syncConfig.length) {
+        return;
+    }
+
+    // Copy
+    for (let item of syncConfig) {
+        if (item.type === 'copy') {
+            CliUtil.doing(`${i18n.copy} '${item.from}' -> '${item.to}'`);
+            await copyDirReadonly(item.from, item.to, !!item.clean, logger);
+            CliUtil.done(true);
+        }
+    }
+
+    // Symlinks
+    await ensureSymlinks(syncConfig.filter(v => v.type === 'symlink').map(v => ({
+        src: v.from,
+        dst: v.to
+    })), console);
 }
 
 export async function copyDirReadonly(src: string, dst: string, clean: boolean, logger?: Logger) {